Move findActivity to utils
